[#3.14] Dermatologist schedule dermatologist appointment

+++ back/src/main/java/com/isa/pharmacies_system/domain/schedule/PharmacistAppointment.java
@@ -30,7 +30,7 @@ public class PharmacistAppointment extends Appointment {
 	private LocalDateTime pharmacistAppointmentStartTime;
 
 	@Column(name="pharmacistAppointmentDuration", unique=false, nullable=false)
-	private double pharmacistAppointmentDuration;
+	private long pharmacistAppointmentDuration;
 
 	@JsonBackReference
 	@ManyToOne(cascade = {CascadeType.ALL}, fetch = FetchType.EAGER)
@@ -52,11 +52,11 @@ public class PharmacistAppointment extends Appointment {
 		this.pharmacistAppointmentStartTime = pharmacistAppointmentStartTime;
 	}
 
-	public double getPharmacistAppointmentDuration() {
+	public long getPharmacistAppointmentDuration() {
 		return pharmacistAppointmentDuration;
 	}
 
-	public void setPharmacistAppointmentDuration(double pharmacistAppointmentDuration) {
+	public void setPharmacistAppointmentDuration(long pharmacistAppointmentDuration) {
 		this.pharmacistAppointmentDuration = pharmacistAppointmentDuration;
 	}
